reduced repeated code in book api with error helper

+++ management/services/book.js
@@ -1,58 +1,65 @@
-
 const bookDao = require("../../dao/bookDao")
+const { required, notFound, alreadyExists } = require("./errorHelper")
 
 // get all books 
-const listAll = async() => {
+const listAll = async () => {
     return bookDao.getAll()
 }
 
 // get book by id 
-const findById = async(id) => {
-    return bookDao.getById(id)
+const findById = async (id) => {
+    const book = await bookDao.getById(id)
+    if (!book) {
+        throw notFound("Book")
+    }
+    return book
 }
 
 // create new book
-const create = async(body) => {
-    if(!body.title){
-        throw new Error ("Title is required")
+const create = async (body) => {
+    const exists = await bookDao.findByTitle(body.title)    // if exist in db
+    if (exists) {
+        throw alreadyExists("Book")
     }
-     const bookExists = await bookDao.findByTitle(body.title.toLowerCase().trim())    // if exist in db
-        if(bookExists){
-            throw new Error ("Book already exists")
-        }
     return bookDao.createRow(body)
 }
 
-
 // put update book
-const replace = async(id, body) => {
-    if(body.title){
-    const bookExists = await bookDao.findByTitle(body.title.toLowerCase().trim());
-    
-        if (bookExists && bookExists.id != id) {
-            throw new Error("Book already exists");
-        }
+const replace = async (id, body) => {
+    const book = await bookDao.getById(id)
+    if (!book) {
+        throw notFound("Book")
+    }
+    const exists = await bookDao.findByTitle(body.title)
+
+    if (exists && exists.id !== Number(id)) {
+        throw alreadyExists("Book")
     }
     return bookDao.replaceRow(id, body)
 }
 
 // patch update (partial) book
-const modify = async(id, body) => {
-
-    if (body.title) {    // if title is update
-    
-            const bookExists = await bookDao.findByTitle(body.title.toLowerCase().trim());
-    
-            if (bookExists && bookExists.id != id) {
-                throw new Error("Book already exists");
-            }
+const modify = async (id, body) => {
+    const book = await bookDao.getById(id)
+    if (!book) {    // if title is update
+        throw notFound("Book")
+    }
+    if (body.title) {
+        const exists = await bookDao.findByTitle(body.title)
+        if (exists && exists.id !== Number(id)) {
+            throw alreadyExists("Book")
         }
+    }
     return bookDao.modifyRow(id, body)
 }
 
 // delete book 
-const remove = async(id) => {
+const remove = async (id) => {
+    const book = await bookDao.getById(id)
+    if (!book) {
+        throw notFound("Book")
+    }
     return bookDao.deleteRow(id)
 }
 
-module.exports = {listAll, findById, create, replace, modify, remove }
+module.exports = { listAll, findById, create, replace, modify, remove }
